docker: Pin to a particular release of Debian Jessie via sha256.
This is done to more explicitly track what version of Debian Jessie is
being used as a base image. It will also ensure that it is properly
updated on the VM even if we forget. This also should help CI and VM
builds stay speedy by using the cache even when there is a newer version
of Debian Jessie. In the long run, we may wish to re-evaluate this
strategy and fix our CI and deployment systems so as to be able to use
the latest version of Debian Jessie with important CVE and other fixes.