Add a tiny alpine based notebook
When doing live demos of JupyterHub  container based deployments,
showmanship is important. And one big aspect of it is 'time taken
from nothing at all to running single-user container'. Image pull
times are a big contributor to this first time demo experience -
and all the current docker-stacks images are too big to give a
'pop'.

This image is optimized for size and demo-worthiness only. Not
to be used for anything else, ideally. Part of the demo can be
configuring the spawner to use a different dockerstack image -
this provides a nice way to explain why it is slow without it
taking away from the initial demo's speed.