Drop the `r` channel
The `r` channel has been considered part of `defaults` since `conda` version `4.3.0`. So it should already be taken into consideration by `conda` installs without having to explicitly add the channel. Further the `conda-forge` channel has incorporated an ever growing, healthy stack of R packages. As such it appears all of the R packages used in this stack now come from `conda-forge` and not `defaults`. Given all of this, it seems safe to drop the `r` channel from explicit addition to the channel list.