Use `MatchSpec` to force `blas` from `conda-forge`
As it appears that `defaults` has added `blas` without a feature, the
`conda` solver appears to be prioritizing the removal of features over
channel priority. As a result, we leverage `MatchSpec` here to pin the
`blas` package to `conda-forge` and the `openblas` feature. This should
be respected by `conda` and thus avoid switching to the incompatible
`blas` package from `defaults`.