Adds config file initialization
Adds the initial creation of the configuration file
jupyter_notebook_config.json to the image. This file
was missing as reported in issue #639 and caused an
error when trying to set a new password via the Web
interface.

Bumps the version of Jupyter Notebook to 5.6

Updates Apache Mirror URL because the current causes build timeouts

The mirror used now (mirrors.ukfast.co.uk) is fast enough and does not
point to an ad site and is already used in a different Dockerfile within
this repository.