Have start.sh run through both user and root parts.
Before, start.sh would run *either* its root or its user part.  If you
were running as uid=0 but wanted to run a user hook before starting,
it can't easily be done.  Also scripts in start-notebook.d and
before-notebook.d could be run as either the user or root, depending
on how things are run.

This is a simple, if unelegant solution, which re-invokes start.sh as
the user, if it was originally invoked as root.