fixed permissions
The USER needs to be set to root at this point because the
fixpermissions that runs at the end is trying to change files that were
copied into places jovyan doesn't have priviledges. I didn't realize
that at first, since I don't normally put files in those places.

Using the $NB_GUID and $NB_UID variables would be better than hardcoding
those values (if someone needs to make them different, for instance),
but wasn't supported in my current version of docker. Apparently this
was fixed three days ago, so it may be possible to use this technique in
the future.