Install nbgitpuller in the base image
[nbgitpuller](https://github.com/jupyterhub/nbgitpuller/) is a very
popular way to distribute content to users, both on mybinder.org
as well as on JupyterHubs. On Binder, this allows for fast launches
where the environment is separated from the content (see
https://discourse.jupyter.org/t/tip-speed-up-binder-launches-by-pulling-github-content-in-a-binder-link-with-nbgitpuller/922).
So by putting nbgitpuller in these images, users can simply pick an
image to launch and then just pull in the content they need, instead
of having to rebuild their own image. It's also extremely popular when
used with JupyterHub, and is included by default in the
[TLJH](https://tljh.jupyter.org) distribution.