base-notebook: stop installing nodejs from conda-forge (#2172)
* base-notebook: stop installing nodejs from conda-forge

nodejs was a conda dependency of jupyterhub, but by installing
jupyterhub-base we no longer need it and could opt to remove it.

By doing this, building base-notebook led to a reported size reduction
from 974MB to 828MB, which is a 146MB / 15% size reduction.

* Update CHANGELOG.md

* Update CHANGELOG.md

* Install jupyterhub-singleuser instead of jupyterhub-base

Note that jupyterhub-base is really whats the foundational need for this
image, where jupyterhub-singleuser the conda-forge package is building on
jupyterhub-base by also adding a depdendency on jupyterlab - but the
jupyterhub-singleuser command is provided by jupyterhub-base conda-forge
package.

* Update CHANGELOG.md

Co-authored-by: Ayaz Salikhov <mathbunnyru@users.noreply.github.com>

---------

Co-authored-by: Ayaz Salikhov <mathbunnyru@users.noreply.github.com>